requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
pyodbc==5.1.0
itsdangerous==2.2.0
python-dotenv==1.0.1
cachetools==5.5.0
qdrant-client==1.11.3
python-docx==1.1.2
duckduckgo-search==6.3.0
PyPDF2==3.0.1
